Created GUI for manipulating objects, concept of physics dispatch, and particle generation for multiple types as well as repaired inheritance of objects.

+++ 3D-fluid-simulator/GraphicEngine/OpenGL_API/VertexBuffer.h
@@ -8,8 +8,8 @@ class VertexBuffer : public Buffer<GL_ARRAY_BUFFER, T>{
 public:
 	using type = T;
 private:
-	const GLuint size = S;
-	GLint _bindingIndex{ -1 };
+	const uint64_t size = S;
+	int32_t _bindingIndex{ -1 };
 public:
 	using type = T;
 	VertexBuffer();
